+++ example/lib/pages/connect.dart
@@ -1,7 +1,3 @@
-//
-//
-//
-
 import 'package:flutter/material.dart';
 import 'package:livekit_client/livekit_client.dart';
 import 'package:shared_preferences/shared_preferences.dart';
