+++ example/lib/theme.dart
@@ -1,7 +1,3 @@
-//
-//
-//
-
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
