+++ example/lib/widgets/no_video.dart
@@ -1,0 +1,20 @@
+import 'package:eva_icons_flutter/eva_icons_flutter.dart';
+import 'package:flutter/material.dart';
+import 'dart:math' as math;
+
+class NoVideoWidget extends StatelessWidget {
+  //
+  const NoVideoWidget({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) => Container(
+        alignment: Alignment.center,
+        child: LayoutBuilder(
+          builder: (ctx, constraints) => Icon(
+            EvaIcons.videoOffOutline,
+            color: Theme.of(ctx).colorScheme.secondary,
+            size: math.min(constraints.maxHeight, constraints.maxWidth) * 0.3,
+          ),
+        ),
+      );
+}
